CURR-43_Corrected Unit Test (#23)
Corrected Unit test and add negative form for methods test

+++ src/main/java/net/ukr/dreamsicle/service/CurrencyService.java
@@ -1,48 +1,68 @@
 package net.ukr.dreamsicle.service;
 
-import lombok.extern.java.Log;
+import lombok.RequiredArgsConstructor;
+import net.ukr.dreamsicle.dto.CurrencyDTO;
+import net.ukr.dreamsicle.dto.CurrencyMapper;
 import net.ukr.dreamsicle.exception.ResourceNotFoundException;
 import net.ukr.dreamsicle.model.Currency;
 import net.ukr.dreamsicle.repository.CurrencyRepositoryDAO;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-@Log
+@RequiredArgsConstructor
 public class CurrencyService {
 
     private final CurrencyRepositoryDAO currencyRepositoryDAO;
+    private final CurrencyMapper currencyMapper;
 
-    @Autowired
-    public CurrencyService(CurrencyRepositoryDAO currencyRepositoryDAO) {
-        this.currencyRepositoryDAO = currencyRepositoryDAO;
+    public List<CurrencyDTO> allCurrencies() {
+        return currencyMapper.toCurrencyDTOs(currencyRepositoryDAO.findAllCurrencies());
     }
 
-    public List<Currency> allCurrenciesData() {
-        return currencyRepositoryDAO.findAllCurrencies();
-    }
-
-    public Currency findCurrencyById(int id) {
+    public CurrencyDTO findCurrencyById(int id) {
         Currency currencyById = currencyRepositoryDAO.findCurrencyById(id);
 
         if (currencyById == null) {
             throw new ResourceNotFoundException();
         }
 
-        return currencyById;
+        return currencyMapper.toCurrencyDto(currencyById);
     }
 
     public void deleteCurrencyById(int id) {
+        Currency currencyById = currencyRepositoryDAO.findCurrencyById(id);
+
+        if (currencyById == null) {
+            throw new ResourceNotFoundException();
+        }
+
         currencyRepositoryDAO.deleteCurrencyById(id);
     }
 
-    public Integer createCurrency(Currency currency) {
-        return currencyRepositoryDAO.createCurrency(currency);
+    public CurrencyDTO createCurrency(CurrencyDTO currencyDTO) {
+        Currency currency = currencyMapper.toCurrency(currencyDTO);
+        Integer id = currencyRepositoryDAO.createCurrency(currency);
+
+        if (id == null) {
+            throw new ResourceNotFoundException();
+        }
+        Currency currencyById = currencyRepositoryDAO.findCurrencyById(id);
+
+        return currencyMapper.toCurrencyDto(currencyById);
     }
 
-    public void updateCurrency(int id, Currency currency) {
+    public CurrencyDTO updateCurrency(int id, CurrencyDTO currencyDTO) {
+        Currency currencyById = currencyRepositoryDAO.findCurrencyById(id);
+
+        if (currencyById == null) {
+            throw new ResourceNotFoundException();
+        }
+
+        Currency currency = currencyMapper.toCurrency(currencyDTO);
+        currency.setId(id);
         currencyRepositoryDAO.updateCurrency(id, currency);
+        return currencyMapper.toCurrencyDto(currencyRepositoryDAO.findCurrencyById(id));
     }
 }
